Add a request spec
