+++ app/views/content/page.html.erb
@@ -1,9 +1,9 @@
-<%= content_for :title, page_title %>
+<%= content_for :title, @page_title %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds-from-desktop">
-    <h1 class="govuk-heading-xl"><%= page_title %></h1>
+    <h1 class="govuk-heading-xl"><%= @page_title %></h1>
 
-    <%= page_content %>
+    <%= @page_content %>
   </div>
 </div>
